backend/requirements.txt
# Certify Intel Backend - Requirements

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Excel
openpyxl==3.1.2

# Web scraping
playwright==1.41.0

# AI/ML
openai==1.10.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0

# Scheduling (for automated scraping)
apscheduler==3.10.4

# Email alerts
aiosmtplib==3.0.1

# PDF Reports
reportlab==4.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Caching (optional)
redis==5.0.1

# SMS / Twilio

# Finance Data
yfinance==0.2.36
pandas==2.1.4
xlsxwriter==3.1.9
matplotlib==3.8.2
seaborn==0.13.1
